test: write test output to 'out' directory
Update XMLGeneratorTest and PDFGeneratorTest to write test outputs to the
'out/test/<testName>' directory, then add 'out/' to the .gitignore. These
files are generated and can change after each test execution, so they should
not be source controlled.

Because the 'out/' directory may not exist when running the test(s), add a
'@BeforeClass' method to create the directories.