mvn: create Maven configuration for project
Replace source controlled library files and Eclipse settings with a Maven
pom.xml defining the project, including how it should be tested and how it
should be built into a single, standalone .jar file. This should generally
make it easier to build and run the program, manage dependencies, and
integrate with a wider variety of IDEs.