MainController.java, ProjectView.java: Creating a save warning after modifying song constants.
In order to enable this, this commit moves the modification tracking
from the ProjectView into the MainController.  Additionally, it cleans
up some tech debt in the Wizard.

Unfortunately, how the views access the saved state and how the
controller updates the project title is still messy due to legacy code.
This can be cleaned up, but would need refactoring outside the scope of
this commit.