MainController.java, MainView.java: Initialize MainView with MainController
Creating a factory-pattern-like method for instantiating the MainController
and MainView together.

This is a workaround for instantiating them safely while they still have
the circular dependency on each other.  Longterm, this circular
dependency should be resolved.

This CL also allows for dependency injection of the MainView into the
MainController, allowing future unit tests to be written against this
controller.