PdfImage.java: replace FootballField with FieldView
Create a new FieldView abstract class that represents a generic drill
FieldView, with the rendering functions simplified from their equivalents in
FootballField by using mathematical helper functions in Point and
transforming the coordinate system to feet before performing the actual
render. Update PdfImage to inherit from FieldView to use the new rendering
code.

The rewritten rendering code adjusts some sizing of things like the
arrowhead on ranks to ideally be slightly more visually appealing and/or
readable. For the most part, though, the drill looks the same as the
FootballField rendering. The main benefit is that the new code allows for
much easier customization based on user preferences (e.g. greyscale coloring
for the PDF).

In future commits, we will also replace the move thumbnails and main drill
editor with classes extending FieldView.