PDFGenerator.java: include sidelines in height
Scale the 'drillHeight' variable based on the total height of the football
field, rather than excluding the sidelines. For now, that doesn't make a
difference in the PDF rendering (since the FootballField class uses its own
measurements), but this will matter in future commits when we render based
on the 'Field' instance dimensions.