FieldView.java: add comment about rank offset error
Somehow, due to either the rank positioning or grid snapping code, all rank
positions are 0.6ft (0.2yd) further back on the field than they should be.
This is visible in the .pnd files, so it's not just a rendering error.

Instead of adding the offset to the render (which could, in the future,
introduce error itself if we fix the rank position issue), add a comment
describing the issue in detail. Separately, we can fix the snapping issue
and, ideally, add versioning to the .pnd file so we can adjust the affected
files automatically when saved in the future.