MainController.java, CommandListView.java: Cleaning up split error logic.
Previously, the split logic in the controller had two different ways of
returning an error - via a return value (which would eventually be
presented in the split dialog) or displaying an error in a dedicated
popup.  Now, the dedicated popup is only used before the split window
opens.  Afterwards, all errors are displayed in the split dialog.